requirements.txt
django>4.2,<4.3
python-decouple
pillow
djangorestframework
markdown       # Markdown support for the browsable API.
django-filter
drf-yasg[validation]
coreapi
djangorestframework-simplejwt
django-mail-templated
djoser
# clean code
black
flake8